refactor: update utility imports to use new shadcn-utils module
