+++ frontend/components/ui/form.tsx
@@ -12,7 +12,7 @@ import {
   type FieldValues,
 } from "react-hook-form"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/lib/shadcn-utils"
 import { Label } from "@/components/ui/label"
 
 const Form = FormProvider
